+++ game/server/hl2/npc_basescanner.cpp
@@ -813,7 +813,7 @@ void CNPC_BaseScanner::PlayFlySound(void)
 
 	float	speed	 = GetCurrentVelocity().Length();
 	float	flVolume = 0.25f + (0.75f*(speed/GetMaxSpeed()));
-	int		iPitch	 = MIN( 255, 80 + (20*(speed/GetMaxSpeed())) );
+	float	iPitch	 = MIN( 255.0f, 80 + (20*(speed/GetMaxSpeed())) );
 
 	//Update our pitch and volume based on our speed
 	controller.SoundChangePitch( m_pEngineSound, iPitch, 0.1f );
